cleanup on subscribe and feedback component
